add delete a superpower

+++ db/schema.rb
@@ -10,15 +10,14 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-
 ActiveRecord::Schema.define(version: 2022_05_24_143321) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
 
   create_table "bookings", force: :cascade do |t|
-    t.integer "user_id", null: false
-    t.integer "superpower_id", null: false
+    t.bigint "user_id", null: false
+    t.bigint "superpower_id", null: false
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
     t.date "starts_at"
@@ -39,7 +38,7 @@ ActiveRecord::Schema.define(version: 2022_05_24_143321) do
     t.string "name"
     t.text "description"
     t.integer "price"
-    t.integer "user_id", null: false
+    t.bigint "user_id", null: false
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
     t.index ["user_id"], name: "index_superpowers_on_user_id"
